+++ src/emulator.cpp
@@ -488,7 +488,7 @@ static int askSave() {
         nullptr,
         "You have unsaved changes to your floppy disk images.\nDo you want to SAVE them?",
         "Save changes",
-        wxYES_NO|wxCANCEL|wxCANCEL_DEFAULT};
+        (long)(wxYES_NO|wxCANCEL|wxCANCEL_DEFAULT)};
 
     return dlg->ShowModal();
 }
